+++ src/main/java/com/braintreegateway/LocalPaymentExpired.java
@@ -1,0 +1,21 @@
+package com.braintreegateway;
+
+import com.braintreegateway.util.NodeWrapper;
+
+public class LocalPaymentExpired {
+    private String paymentId;
+    private String paymentContextId;
+
+    public LocalPaymentExpired(NodeWrapper node) {
+        this.paymentId = node.findString("payment-id");
+        this.paymentContextId = node.findString("payment-context-id");
+    }
+
+    public String getPaymentId() {
+        return paymentId;
+    }
+
+    public String getPaymentContextId() {
+        return paymentContextId;
+    }
+ }